temporarily remove PhantomJSDriverSupplier
When you declare in the pom, the dep, it replaces selenium-api 2.53.0 (declared in my pom) with whatever version phantomjsdriver depends on.
I can't seem to be able to force the latest.
The issue seems to be that selenium-api is not declared as a dependency of selenium-java, but it's a dependency of selenium-parent,
therefore I guess that maven's dependencyManagement section is not considered in the equation.

TODO: put PhantomJSDriverSupplier back in